fix: address Nim 2.2 JS backend compatibility issues
Fix three Nim 2.2 JS backend regressions:

1. Method dispatch: Declare onUpdatedEvents base method on Component
   instead of EventLogComponent, so Nim 2.2 generates the override
   correctly on the JS backend.

2. Bool serialization: Nim 2.2 toJs serializes array[enum, bool] with
   numeric values (1/0) instead of JSON booleans (true/false). The Rust
   db-backend selectedKinds deserializer now handles both via a
   value_as_truthy helper, fixing the event log showing no data.

3. BigInt/overflow: Nim 2.2 maps int64 to JS BigInt, causing Cannot mix
   BigInt and other types errors. Changed renderRRTicksLine to use int
   parameters with float arithmetic, avoiding both BigInt mixing and
   32-bit overflow.

Additional defensive fixes:
- Bounds check response[0] access in onUpdatedEvents (empty response guard)
- Wrap flow.nim getCurrentStepCount and renderFlowLines in try-except
  IndexDefect to handle out-of-bounds loop/step access gracefully

+++ src/frontend/ui/trace_log.nim
@@ -54,7 +54,7 @@ proc refreshTraces(self: TraceLogComponent) =
             js{
               className: cstring"direct-location-rr-ticks",
               data: cstring"rrTicks",
-              render: proc(rrTicks: int64): cstring =
+              render: proc(rrTicks: int): cstring =
                 renderRRTicksLine(rrTicks, self.data.minRRTicks, self.data.maxRRTicks, "event-rr-ticks-line")
             },
             js{
